Split ReferralsServiceDelegate into another file
