+++ browser/brave_referrals/sources.gni
@@ -1,0 +1,20 @@
+# Copyright (c) 2023 The Brave Authors. All rights reserved.
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at https://mozilla.org/MPL/2.0/.
+
+import("//brave/components/brave_referrals/buildflags/buildflags.gni")
+
+brave_browser_brave_referrals_sources = []
+brave_browser_brave_referrals_deps = []
+
+if (enable_brave_referrals) {
+  brave_browser_brave_referrals_sources += [
+    "//brave/browser/brave_referrals/referrals_service_delegate.cc",
+    "//brave/browser/brave_referrals/referrals_service_delegate.h",
+  ]
+  brave_browser_brave_referrals_deps += [
+    "//base",
+    "//chrome/common",
+  ]
+}
