+++ browser/sources.gni
@@ -7,6 +7,7 @@ import("//brave/browser/autocomplete/sources.gni")
 import("//brave/browser/brave_adaptive_captcha/sources.gni")
 import("//brave/browser/brave_ads/sources.gni")
 import("//brave/browser/brave_news/sources.gni")
+import("//brave/browser/brave_referrals/sources.gni")
 import("//brave/browser/brave_rewards/sources.gni")
 import("//brave/browser/brave_shields/sources.gni")
 import("//brave/browser/brave_stats/sources.gni")
@@ -420,6 +421,7 @@ brave_chrome_browser_sources += brave_browser_themes_sources
 brave_chrome_browser_sources += brave_browser_web_discovery_sources
 brave_chrome_browser_sources += brave_browser_url_sanitizer_sources
 brave_chrome_browser_sources += brave_browser_sharing_hub_sources
+brave_chrome_browser_sources += brave_browser_brave_referrals_sources
 
 brave_chrome_browser_deps += brave_browser_autocomplete_deps
 brave_chrome_browser_deps += brave_browser_brave_adaptive_captcha_deps
@@ -453,3 +455,4 @@ brave_chrome_browser_deps += brave_browser_web_discovery_deps
 brave_chrome_browser_deps += brave_chromium_src_chrome_browser_deps
 brave_chrome_browser_deps += brave_chromium_src_chrome_browser_prefs_deps
 brave_chrome_browser_deps += brave_browser_url_sanitizer_deps
+brave_chrome_browser_deps += brave_browser_brave_referrals_deps
